requirements.txt
# Core ML frameworks
torch==2.10.0
torchvision==0.25.0
transformers==5.2.0

# Image processing
pillow==12.1.1

# Data & visualisation
numpy==2.4.2
matplotlib==3.10.8
scikit-learn==1.8.0

# Utilities
timm==1.0.25